Crate hotg_rune_core

Modules

capabilities

outputs

reflect

A simple reflection system.

Structs

BufWriter

InvalidConversionError

SerializableRecord

A serializable version of log::Record.

Shape

A tensor’s shape.

Tensor

A multidimensional array with copy-on-write semantics.

TensorView

An immutable view into a Tensor with a particular rank (number of dimensions).

TensorViewMut

A mutable view into a Tensor with a particular rank (number of dimensions).

Enums

PixelFormat

PixelFormatConversionError

Type

Value

A dynamically typed value that may be passed back and forth across the runtime.

Constants

TFLITE_MIMETYPE

The mimetype used for a TensorFlow Lite model.

Traits

AsType

A Rust primitive which has a corresponding Type and can be converted to or from a Value.

HasOutputs

Sink

A consumer of data.

Source

A stream of data, typically something like a random number generator or sensor.

TensorList

A helper trait which lets us get the shape from a tuple of different tensors.

TensorListMut

A set of tensors that can be mutated.